move module into its own package (fix #62)
