+++ main.go
@@ -1,11 +1,14 @@
 package main
 
-import "github.com/lyft/protoc-gen-star"
+import (
+	"github.com/lyft/protoc-gen-star"
+	"github.com/lyft/protoc-gen-validate/module"
+)
 
 func main() {
 	pgs.
 		Init(pgs.DebugEnv("DEBUG_PGV")).
-		RegisterModule(Validator()).
+		RegisterModule(module.Validator()).
 		RegisterPostProcessor(pgs.GoFmt()).
 		Render()
 }
