+++ module/checker.go
@@ -1,4 +1,4 @@
-package main
+package module
 
 import (
 	"reflect"
